+++ armada_backend/consul_config.py
@@ -46,6 +46,7 @@ def get_consul_config(consul_mode, ship_ips, datacenter, ship_external_ip):
         {'type': 'keyprefix', 'prefix': 'dockyard/', 'handler': save_runtime_settings_cmd},
         {'type': 'nodes', 'handler': save_runtime_settings_cmd},
         {'type': 'checks', 'handler': save_running_containers_cmd},
+        {'type': 'keyprefix', 'prefix': 'ships/', 'handler': save_running_containers_cmd},
     ]
 
     return json.dumps(config, sort_keys=True, indent=4)